Add scoreboard graphics and flashing powerups

+++ Cap-Man/Animation.cpp
@@ -6,7 +6,7 @@ Animation::Animation()
 
 Animation::Animation(int spriteIntervalMillis)
 	: mSpriteIntervalMillis(spriteIntervalMillis)
-	, mTimeSpentBetweenSpriteChange(0)
+	, mTimeSpentBetweenSpriteChange(0.0f)
 	, mCurrentSpriteIndex(0) {
 }
 
@@ -14,16 +14,16 @@ Animation::~Animation() {
 }
 
 void Animation::reset() {
-	mTimeSpentBetweenSpriteChange = 0;
+	mTimeSpentBetweenSpriteChange = 0.0f;
 	mCurrentSpriteIndex = 0;
 }
 
 void Animation::update(float delta) {
-	mTimeSpentBetweenSpriteChange += static_cast<int>(delta * 1000.0f /* millis */);
+	mTimeSpentBetweenSpriteChange += delta * 1000.0f /* millis */;
 
 	if (mTimeSpentBetweenSpriteChange >= mSpriteIntervalMillis) {
 		mCurrentSpriteIndex = (mCurrentSpriteIndex + 1) % mSprites.size();
-		mTimeSpentBetweenSpriteChange = 0;
+		mTimeSpentBetweenSpriteChange = 0.0f;
 	}
 }
